+++ shared-module/displayio/__init__.h
@@ -42,7 +42,7 @@
 #include "shared-bindings/rgbmatrix/RGBMatrix.h"
 #endif
 #if CIRCUITPY_IS31FL3741
-#include "shared-bindings/is31fl3741/is31fl3741.h"
+#include "shared-bindings/is31fl3741/IS31FL3741.h"
 #endif
 #if CIRCUITPY_SHARPDISPLAY
 #include "shared-module/sharpdisplay/SharpMemoryFramebuffer.h"
